pyproject.toml
[project]
name = "amgec26"
version = "5.1.0"
description = "AMGEC 2026 CONFERENCE WEBSITE"
author = "NAJLA BEN HASSINE - LaNSER Webmaster"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5==2025.6",
    "django==5.2.7",
    "django-admin-filter==0.7",
    "django-admin-interface==0.30.1",
    "django-bootstrap-icons==0.9.0",
    "django-bootstrap5==25.2",
    "django-clearcache==1.2.1",
    "django-countries==7.6.1",
    "django-crispy-forms==2.4",
    "django-filter==25.2",
    "django-import-export==4.3.12",
    "fontawesome-free==5.15.4",
    "gunicorn==23.0.0",
    "mysql-connector-python==9.4.0",
    "mysqlclient==2.2.7",
    "openpyxl==3.1.5",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.3",
]
